Refactor channel index types from int to size_t across multiple processors and related components
- Updated channel index types in compressor, lookahead limiter, noise gate, and race processors to use size_t instead of int for improved type safety and consistency.
- Adjusted validation checks to ensure channel indices are non-negative and within valid ranges.
- Modified memory allocation and copying logic to accommodate size_t types for channel indices.
- Updated related tests to reflect the change in channel index types, ensuring compatibility and correctness.
- Refactored spectrum handling functions to use size_t for channel parameters, enhancing type safety in spectrum analysis.

+++ Audio/audio_chunk.h
@@ -142,9 +142,9 @@ void round_robin_chunk_pool_free(round_robin_chunk_pool_t* pool);
  * @param out_sum Output buffer to write the summed samples.
  * @param frames Number of frames to sum.
  */
-void audio_chunk_sum_channels(const audio_chunk_t* chunk, const int* channels,
-                              size_t channels_count, double* out_sum,
-                              size_t frames);
+void audio_chunk_sum_channels(const audio_chunk_t* chunk,
+                              const size_t* channels, size_t channels_count,
+                              double* out_sum, size_t frames);
 
 /**
  * @brief Applies sample-by-sample linear gain to multiple channels in a chunk.
@@ -158,7 +158,7 @@ void audio_chunk_sum_channels(const audio_chunk_t* chunk, const int* channels,
  * @param gain_multipliers Array of linear gain multipliers (one per frame).
  * @param frames Number of frames to process.
  */
-void audio_chunk_apply_gain(audio_chunk_t* chunk, const int* channels,
+void audio_chunk_apply_gain(audio_chunk_t* chunk, const size_t* channels,
                             size_t channels_count,
                             const double* gain_multipliers, size_t frames);
