+++ Server/websocket_server.c
@@ -549,10 +549,9 @@ static void* server_thread_func(void* arg) {
             cdsp_spectrum_side_t side_val = session->spectrum_is_capture
                                                 ? CDSP_SPECTRUM_SIDE_CAPTURE
                                                 : CDSP_SPECTRUM_SIDE_PLAYBACK;
-            const uint32_t* chan_ptr =
-                (session->spectrum_channel == (uint32_t)-1)
-                    ? NULL
-                    : &session->spectrum_channel;
+            const size_t* chan_ptr = (session->spectrum_channel == (size_t)-1)
+                                         ? NULL
+                                         : &session->spectrum_channel;
             bool spec_ok =
                 (p_freqs && p_mags && server && server->engine) &&
                 cdsp_get_spectrum(server->engine, side_val, chan_ptr,
